+++ lib/pack_game/levels_hud/controls/widgets/level_actions_row.dart
@@ -6,70 +6,259 @@ import 'package:wbw_core/wbw_core.dart';
 import 'package:wbw_design_core/wbw_design_core.dart';
 import 'package:word_by_word_game/generated/l10n.dart';
 import 'package:word_by_word_game/pack_core/global_states/global_states.dart';
+import 'package:word_by_word_game/pack_game/levels_hud/controls/widgets/player_swither_bar.dart';
+import 'package:word_by_word_game/pack_game/levels_hud/controls/widgets/word_actions_buttons.dart';
 import 'package:word_by_word_game/pack_game/levels_hud/controls/widgets/word_composition_row.dart';
 import 'package:word_by_word_game/pack_game/mechanics/mechanics.dart';
 
-class UILevelActionsRow extends HookWidget {
-  const UILevelActionsRow({
-    this.leftTopBuilder,
-    this.rightTopBuilder,
-    final Key? key,
-  }) : super(key: key);
-  final WidgetBuilder? leftTopBuilder;
-  final WidgetBuilder? rightTopBuilder;
-  static const kIsCookingEnabled = kDebugMode;
+class UIMobileLevelActionsRow extends HookWidget {
+  const UIMobileLevelActionsRow({super.key});
+
   @override
   Widget build(final BuildContext context) {
     final widgetState = context.read<WordCompositionState>();
     final uiTheme = UiTheme.of(context);
-    final mechanicsCollection = context.read<MechanicsCollection>();
-    return UICenterFrame(
-      onIdea: null,
-      onPause: widgetState.onPause,
-      leftTopBuilder: leftTopBuilder,
-      rightTopBuilder: rightTopBuilder,
-      textFieldBuilder: (final context) {
-        return BlocBuilder<LevelBloc, LevelBlocState>(
-          buildWhen: LevelBloc.useCheckStateEqualityBuilder(
-            checkLiveState: (final previous, final current) =>
-                previous.actionMultiplier != current.actionMultiplier ||
-                previous.actionType != current.actionType,
+    final theme = Theme.of(context);
+    final textTheme = theme.textTheme;
+
+    return BlocBuilder<LevelBloc, LevelBlocState>(
+      buildWhen: LevelBloc.useCheckStateEqualityBuilder(
+        checkLiveState: (final previous, final current) =>
+            previous.actionMultiplier != current.actionMultiplier ||
+            previous.actionType != current.actionType,
+      ),
+      builder: (final context, final levelState) {
+        if (levelState is! LiveLevelBlocState) {
+          return const SizedBox();
+        }
+        return Padding(
+          padding: EdgeInsets.symmetric(horizontal: uiTheme.spacing.medium),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  const Expanded(child: UIMobilePlayerName()),
+                  UiPauseIconButton(onPressed: widgetState.onPause),
+                  const Expanded(
+                    child: Align(
+                      alignment: Alignment.centerRight,
+                      child: UIMobilePlayerScore(),
+                    ),
+                  ),
+                ],
+              ),
+              uiTheme.verticalBoxes.large,
+              Row(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  const Expanded(
+                    child: UiActionFrame(),
+                  ),
+                  uiTheme.horizontalBoxes.large,
+                  const UiMobileEffectFrame()
+                ],
+              )
+            ],
           ),
-          builder: (final context, final levelState) {
-            if (levelState is! LiveLevelBlocState) {
-              return const SizedBox();
-            }
-            return Column(
-              children: [
-                Text(S.of(context).selectAction),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    if (kIsCookingEnabled)
-                      UILevelActionChip(
-                        baseText: S.of(context).cookFood,
-                        levelState: levelState,
-                        scoreBuilder: () => mechanicsCollection.score
-                            .getScoreForCookFoodByModifier(
-                          multiplier: levelState.actionMultiplier,
-                        ),
-                        type: LevelPlayerActionType.cookFood,
-                      ),
-                    uiTheme.horizontalBoxes.medium,
-                    UILevelActionChip(
-                      baseText: S.of(context).refuelStorage,
-                      levelState: levelState,
-                      scoreBuilder: () => mechanicsCollection.score
-                          .getScoreForRefuelStorageByModifier(
-                        multiplier: levelState.actionMultiplier,
-                      ),
-                      type: LevelPlayerActionType.refuelStorage,
+        );
+      },
+    );
+  }
+}
+
+class UIDesktopLevelActionsRow extends HookWidget {
+  const UIDesktopLevelActionsRow({super.key});
+  @override
+  Widget build(final BuildContext context) {
+    final widgetState = context.read<WordCompositionState>();
+    final uiTheme = UiTheme.of(context);
+    final theme = Theme.of(context);
+    final textTheme = theme.textTheme;
+
+    return BlocBuilder<LevelBloc, LevelBlocState>(
+      buildWhen: LevelBloc.useCheckStateEqualityBuilder(
+        checkLiveState: (final previous, final current) =>
+            previous.actionMultiplier != current.actionMultiplier ||
+            previous.actionType != current.actionType,
+      ),
+      builder: (final context, final levelState) {
+        if (levelState is! LiveLevelBlocState) {
+          return const SizedBox();
+        }
+        return Container(
+          width: 300,
+          padding: EdgeInsets.symmetric(horizontal: uiTheme.spacing.medium),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  const UIMobilePlayerName(),
+                  UiPauseIconButton(onPressed: widgetState.onPause),
+                  const UIMobilePlayerScore(),
+                ],
+              ),
+              uiTheme.verticalBoxes.medium,
+              Container(
+                padding: EdgeInsets.all(uiTheme.spacing.medium),
+                decoration: BoxDecoration(
+                  border: Border.all(),
+                  borderRadius: BorderRadius.all(uiTheme.circularRadius.small),
+                ),
+                child: Row(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: const [
+                    Flexible(
+                      child: UiActionFrame(),
                     ),
                   ],
                 ),
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+}
+
+class UiActionFrame extends StatelessWidget {
+  const UiActionFrame({super.key});
+  static const kIsCookingEnabled = kDebugMode;
+
+  @override
+  Widget build(final BuildContext context) {
+    final uiTheme = UiTheme.of(context);
+    final theme = Theme.of(context);
+    final textTheme = theme.textTheme;
+    return BlocBuilder<LevelBloc, LevelBlocState>(
+      buildWhen: LevelBloc.useCheckStateEqualityBuilder(
+        checkLiveState: (final previous, final current) =>
+            previous.actionMultiplier != current.actionMultiplier ||
+            previous.actionType != current.actionType,
+      ),
+      builder: (final context, final levelState) {
+        if (levelState is! LiveLevelBlocState) {
+          return const SizedBox();
+        }
+        return Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Text(S.of(context).selectAction),
+            uiTheme.verticalBoxes.medium,
+            Wrap(
+              spacing: uiTheme.spacing.medium,
+              runSpacing: uiTheme.spacing.medium,
+              children: [
+                if (kIsCookingEnabled)
+                  UILevelActionChip(
+                    baseText: S.of(context).cookFood,
+                    levelState: levelState,
+                    type: LevelPlayerActionType.cookFood,
+                  ),
+                UILevelActionChip(
+                  baseText: S.of(context).refuelStorage,
+                  levelState: levelState,
+                  type: LevelPlayerActionType.refuelStorage,
+                ),
               ],
-            );
-          },
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
+
+class UiDesktopEffectFrame extends StatelessWidget {
+  const UiDesktopEffectFrame({super.key});
+
+  @override
+  Widget build(final BuildContext context) {
+    final uiTheme = UiTheme.of(context);
+    final theme = Theme.of(context);
+    final textTheme = theme.textTheme;
+    return ConstrainedBox(
+      constraints: const BoxConstraints(minWidth: 200),
+      child: BlocBuilder<LevelBloc, LevelBlocState>(
+        buildWhen: LevelBloc.useCheckStateEqualityBuilder(
+          checkLiveState: (final previous, final current) =>
+              previous.actionMultiplier != current.actionMultiplier ||
+              previous.actionType != current.actionType,
+        ),
+        builder: (final context, final levelState) {
+          if (levelState is! LiveLevelBlocState) {
+            return const SizedBox();
+          }
+          return Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                S.of(context).effect,
+                style: textTheme.titleSmall,
+              ),
+              uiTheme.verticalBoxes.small,
+              Row(
+                children: LevelActionMultiplierType.values.map(
+                  (final type) {
+                    return UILevelActionMultiplierChip(
+                      levelState: levelState,
+                      type: type,
+                    );
+                  },
+                ).toList(),
+              ),
+            ],
+          );
+        },
+      ),
+    );
+  }
+}
+
+class UiMobileEffectFrame extends StatelessWidget {
+  const UiMobileEffectFrame({super.key});
+
+  @override
+  Widget build(final BuildContext context) {
+    final uiTheme = UiTheme.of(context);
+    final theme = Theme.of(context);
+    final textTheme = theme.textTheme;
+    return BlocBuilder<LevelBloc, LevelBlocState>(
+      buildWhen: LevelBloc.useCheckStateEqualityBuilder(
+        checkLiveState: (final previous, final current) =>
+            previous.actionMultiplier != current.actionMultiplier ||
+            previous.actionType != current.actionType,
+      ),
+      builder: (final context, final levelState) {
+        if (levelState is! LiveLevelBlocState) {
+          return const SizedBox();
+        }
+        return ConstrainedBox(
+          constraints: const BoxConstraints(minWidth: 80),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                S.of(context).effect,
+                style: textTheme.titleSmall,
+              ),
+              uiTheme.verticalBoxes.small,
+              ...LevelActionMultiplierType.values.map(
+                (final type) {
+                  return UILevelActionMultiplierChip(
+                    levelState: levelState,
+                    type: type,
+                  );
+                },
+              ),
+            ],
+          ),
         );
       },
     );
@@ -80,29 +269,63 @@ class UILevelActionChip extends StatelessWidget {
   const UILevelActionChip({
     required this.levelState,
     required this.baseText,
-    required this.scoreBuilder,
     required this.type,
     super.key,
   });
   final LevelPlayerActionType type;
   final LiveLevelBlocState levelState;
-  final ValueGetter<ScoreModel> scoreBuilder;
   final String baseText;
   @override
   Widget build(final BuildContext context) {
-    String cost = '';
     final widgetState = context.read<WordCompositionState>();
     final isSelected = levelState.actionType == type;
-    if (isSelected) {
-      final score = scoreBuilder();
-      cost = ' -${score.value.toInt()}';
-    }
     return ChoiceChip(
       selected: isSelected,
       onSelected: (final _) => widgetState.onSelectActionType(type),
       label: Text(
-        '$baseText'
-        '${isSelected ? cost : ""}',
+        baseText,
+      ),
+    );
+  }
+}
+
+class UILevelActionMultiplierChip extends StatelessWidget {
+  const UILevelActionMultiplierChip({
+    required this.levelState,
+    required this.type,
+    super.key,
+  });
+  final LevelActionMultiplierType type;
+  final LiveLevelBlocState levelState;
+  @override
+  Widget build(final BuildContext context) {
+    final widgetState = context.read<WordCompositionState>();
+    final mechanics = context.read<MechanicsCollection>();
+
+    final player =
+        context.select<LevelPlayersBloc, PlayerProfileModel>((final bloc) {
+      final liveState = bloc.getLiveState();
+      return liveState.currentPlayer;
+    });
+    final applyingScore = mechanics.score
+        .getScoreForRefuelStorageByModifier(multiplier: type)
+        .value
+        .toInt();
+    final isAllowedToUse = mechanics.score.checkPlayerAbilityToUseScore(
+      player: player,
+      score: applyingScore,
+    );
+
+    final bool isSelected =
+        isAllowedToUse && levelState.actionMultiplier == type;
+    return Tooltip(
+      message: isAllowedToUse ? '' : S.of(context).youDontHaveEnoughPoints,
+      child: ChoiceChip(
+        selected: isSelected,
+        onSelected: !isAllowedToUse
+            ? null
+            : (final _) => widgetState.onSelectActionMultiplier(type),
+        label: Text('$applyingScore'),
       ),
     );
   }