+++ lib/pack_game/settings/widgets/locale_switcher_button.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:intl/intl.dart' as intl;
 import 'package:life_hooks/life_hooks.dart';
 import 'package:provider/provider.dart';
 import 'package:wbw_core/wbw_core.dart';
@@ -17,18 +16,22 @@ class LocaleSwitcherButton extends HookWidget {
   Widget build(final BuildContext context) {
     final state = _useLocaleSwitcherButtonState(read: context.read);
 
-    return DropdownButton<Locale>(
+    return DropdownButton<Locale?>(
       value: state.locale,
       onChanged: state.onUpdateLocale,
-      items: namedLocalesMap.values
-          .map(
-            (final e) => DropdownMenuItem<Locale>(
-              value: e.locale,
-              key: ValueKey(e.code),
-              child: SettingsText(text: e.name),
-            ),
-          )
-          .toList(),
+      items: [
+        DropdownMenuItem<Locale>(
+          key: const ValueKey('null'),
+          child: SettingsText(text: state.systemLocale.getValue()),
+        ),
+        ...namedLocalesMap.values.map(
+          (final e) => DropdownMenuItem<Locale>(
+            value: e.locale,
+            key: ValueKey(e.code),
+            child: SettingsText(text: e.name),
+          ),
+        ),
+      ],
     );
   }
 }